Tratamento de exceções dos métodos delete e update

+++ src/main/java/com/educandoweb/course/service/UserService.java
@@ -4,11 +4,15 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.repositories.UserRepository;
+import com.educandoweb.course.service.exceptions.DataBaseException;
 import com.educandoweb.course.service.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class UserService {
@@ -31,14 +35,25 @@ public class UserService {
 	}
 
 	public void delete(Long id) {
-		Repository.deleteById(id);
-
+		if (!Repository.existsById(id)) {
+			throw new ResourceNotFoundException(id);
+		}
+		try {
+			Repository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DataBaseException(e.getMessage());
+		}
 	}
 	
 	public User update(Long id, User obj) {
-		User entity = Repository.getReferenceById(id);
-		updateData(entity, obj);
-		return Repository.save(entity);
+		try {
+			User entity = Repository.getReferenceById(id);
+			updateData(entity, obj);
+			return Repository.save(entity);
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException(id);
+		}
+
 	}
 
 	private void updateData(User entity, User obj) {
